OBJC-C compatibility
To be compatible for OBJC we need to set the protocole as @objc and use
the keyword NSObjectProtocol instead of class. also OBJC does not like
arrays containing CGColor as they are not object in objc so we need to
use UIColor as mush as possible for every method call that are public
API calls.